Remove Lombok dependency
And "de-lombok" PriceData, which was the only class using Lombok. The
overhead of needing the Lombok IDEA plugin and annotation processor
enabled is just not worth the few lines of code saved in this one class.

Now the project is as vanilla Java 8 as it gets, and therefore that much
easier to get started with.