Upgrade to cache-aware version of heroku-buildpack-tor
In response to my comment at
https://github.com/apg/heroku-buildpack-tor/issues/3#issuecomment-324105065,
@hernanex3 has been kind enough to implement support for Heroku's
CACHE_DIR in his fork of the Tor buildpack. This commit upgrades to that
version with the following commands:

    heroku buildpacks:remove https://github.com/hernanex3/heroku-buildpack-tor.git
    heroku buildpacks:add https://github.com/hernanex3/heroku-buildpack-tor.git#c766cb6d39f000603dc048f3f82d13b76c4fbf31

Note that I am now pinning the buildpack URL to the commit in question
(hernanex3/heroku-buildpack-tor@c766cb6d) to ensure there is no chance
of future changes to this buildpack breaking the documented approach to
deploying on Heroku.