Introduce ExchangeRateProvider SPI
This change normalize all existing exchange rate providers to conform to
this interface, including BitcoinAverage, which previously was the
special case, having getLocal and getGlobal methods instead of a single
request method. This is now no longer a special case, and simply treated
as two separate price providers that share a common base class.