Rename CachingExchangeRateData#{requestIntervalMs => ttl}
'ttl' (time to live) reveals the intention of this property more clearly
than requestIntervalMs, for it indeed represents how long exchange rate
data is supposed to live before being refreshed. Using an acronym here
vs spelling it out as ttl/TTL is quite well known, and pretty obvious in
the context of this explicitly-named caching component.