Remove ExchangeRateService#removeOutdatedPrices
This is no longer necessary now that the data structure that stores this
information is no longer long-lived. It is recreated on each call to
getAllMarketData, which in turn returns the latest cached data from
underlying providers. This data is never older than the individual
providers' TTLs, so there is never 30-minute old data to purge, which is
what this method was for; it's just obsolete now.