Improve error handling
Make sure that failures at pricenode startup kill the JVM, and that
failures on scheduled requests do not.

The Timer objects used to schedule requests are backed by non-daemon
threads by default. This means that they keep the JVM alive until they
are explicitly shut down. Generally, this is what we want: a Pricenode
should run indefinitely once started, periodically fetching and caching
new price information and returning it to Bisq clients on request.

However, if there is a failure at startup time, for example because of
a missing or invalid BitcoinAverage keypair, the error should be
reported and the JVM should halt immediately. To make this happen, this
commit introduces a Pricenode#stop method, which is called any time an
exception propagates all the way up to the main method. This #stop
method ultimately calls Timer#cancel on all underlying timer objects,
shutting down these non-daemon threads and allowing the JVM to halt as
expected.

In the case of errors that occur after startup, for example one of our
exchange rate providers like BitcoinAverage having an outage, we most
certainly do *not* want our Pricenode to halt, and for this reason we
catch any such exceptions within the scheduled call logic and simply
report it to the console; we do not rethrow the exception and allow it
to propagate up the stack and kill the JVM. In the future, we might want
to introduce exponential backoffs, circuit-breaker style functionality
and alerting mechanisms to make operators aware of these kinds of
failures, but for now the goal is to make sure that basic error handling
works as it should, i.e. allowing Pricenodes to continue operating
indefinitely even in the face of an unreliable network and unreliable
provider endpoints.

Throughout the codebase, explicitly 'throws IOException' clauses have
been removed in favor of catching IOException, wrapping and rethrowing
them as UncheckedIOException. This change facilitates all those
mentioned above.

Note that this change is not strictly a refactoring as it does change
the behavior of handling errors that occur at startup time, but it is
still a refactoring in spirit as it does not affect the behavior of the
application at runtime.