Remove legacy log initialization logic
This change is not strictly a refactoring as it means that logs are no
longer written to $HOME/provider, but it is assumed that this is not a
problem for operators, and not critical to the operation of their nodes.

It does not change the runtime behavior of a pricenode in any way. It is
assumed that this kind of log initialization logic was more or less
copied-and-pasted from the main bisq-network/exchange log initialization
logic, and not terribly important that it worked this way. It is being
removed here in anticipation of introducing Spring Boot, which will have
it's own log management logic that 'just works' in a well-known way.