Migrate runtime and configuration to Spring
This change introduces the latest version of Spring Boot to:

 - Replace the Spark web framework and its underlying use of Jetty with
   Spring MVC and Spring Boot's default embedded Tomcat runtime

 - Replace our own custom Environment with Spring's Environment

 - Replace our own ServiceLoader based instantiation with Spring's
   annotation-based dependency injection

Along the way, this change also removes the use of Gradle's
'application' plugin (and its generated scripts) altogether in favor of
Spring Boot's Gradle plugin and the fat executable jar it creates. All
README instructions, etc have been updated accordingly.

Again, this is all pure refactoring, no behavior changes to the running
app. Generally, these changes make the application simpler, smaller and
easier to deal with.

This commit is the beginning of several to follow, which will migrate
other aspects of the application to use Spring infrastructure rather
than rolling our own.