Fix incompatible JSON serialization in /getAllMarketPrices
Previous changes in this branch modified the format of the 'data'
element returned by /getAllMarketPrices such that its value was a map of
maps, rather than an array of maps, as it should be. This change
restores the shape of the data to the latter.

This change also fixes the naming of the keys in the array of maps to
confirm to the existing format, e.g. 'timestampSec' instead of
'timestamp', and 'currencyCode' instead of 'currency'. This is done with
explicit @JsonProperty annotations such that (a) the naming of the
members in the code can remain as-is, and (b) so that any future changes
to those members' names do not cause a similar incompatibility to arise.