Extract common controller logging to WebConfig
This class consolidates previously duplicated logic for logging
incoming requests. An alternative approach would have been to introduce
a common base class for all controllers with a @ModelAttribute method.
This might be a better approach in the end, as it ends up clearly
logging which controller is going to handle the request. Here, the
logging category is always bisq.price.WebConfig, which does not make it
obvious which controller will actually handle the request.