Introduce bisq.price.PriceProvider base class
This generic abstract base class eliminates the need for
AbstractExchangeRateProvider and CachingExchangeRateProvider and will
(in subsequent commits) serve as the base class for FeeEstimationProvider
on the mining side as well.

The net effect is that there will be a single place where scheduled
caching logic is defined.

Note also the new onRefresh hook and it's overridden implementation in
ExchangeRateProvider for logging purposes.