Remove use of io.bisq.common.util.MathUtils
The #roundDoubleToLong functionality is not necessary here. Values
coming in from bitcoinfees are always some number of satoshis followed
by a single zero decimal value, e.g. 0.0, 10.0, 20.0, etc. This means
that the standard Double#longValue() is sufficient.

This is being done toward the larger goal of removing completely our
dependency on the bisq-exchange:core library.