Derive version from latest git tag
This change replaces the previous custom version management with
Palantir's git-version Gradle plugin, which derives the application
version based on the output of `git describe`. See the git-describe
manpage for details about how this works.

The benefit of this approach is that we never need to manually change
the version string in a text file. Git is the canonical source of
version information, as it should be. And thanks to the way `git
describe` works, our version strings will always include precise
information about exactly what commit the application was built from.
For example, if the application was built from the v0.7.1 tag, the
version string will be "v0.7.1". If, however, the application was built
from a revision 3 commits after the v0.7.1 tag, the version string will
be "v0.7.1-3-g1234567", where "1234567" is the short version of the git
commit hash from which the application was built. Again, see the
git-describe manpage for full details.

This change also enables the default set of Spring Boot actuator
endpoints, and configures application.properties such that the
application version is exposed via the /info endpoint. This will
eventually replace the need for VersionController altogether, but it
(and its /getVersion) endpoint has been left in place for compatibility
with existing monitoring solutions.

We are no longer writing Implementation-Title and Implementation-Version
metadata in the jar manifest, as we are no longer reading it anywhere.
It wouldn't hurt to leave it, but removing it simplifies the build
script and removes confusion about the source of version information in
the application.

Finally, this change modifies the startup banner to include the version
string, but in order to do so, it was necessary to escape all
backslashes in banner.txt to avoid exceptions being thrown from Groovy's
SimpleTemplateEngine.