Configure Heroku to handle Git-based app version
Even though Heroku is fundamentally Git-based, it does not actually keep
the .git directory in place in the final directory where the application
is launched.

This change introduces a custom Heroku buildpack that clones a target
repository at a given revision / branch and drops its .git directory
into the Heroku application directory such that Gradle-based Git version
management (introduced in previous commits) can work as intended.

It is important that the value of GIT_CLONE_URL and GIT_CLONE_SHA are
the same Git remote that causes automatic deployments to Heroku. Note
that it is still fine to push directly to the Heroku remote, e.g. with
`git push heroku some-local-branch:master`, but the version string will
always show up as 'dirty' when you do this.