Run PriceProvider#timer instances as daemon threads
This ensures that when an exception is thrown at Spring container
startup time, the application will actually exit, as opposed to staying
running after the exception has been thrown.

This change is closely related to similar changes made in
c0eb50a78145604a1c2b237893771cca0f149608 prior to the introduction of
Spring, and acheives the same effect: failing fast and actually shutting
down the JVM when there is an exception at startup time.