Revert to simple version management
This change largely reverts Git-based versioning introduced in commit
7de26d7470dd2ec1e45bc37fe06873a619816b22, as in the end it introduced
too much complexity and special handling for the Heroku case. It would
probably have caused other problems for other hosting providers and
deployment scenarios as well.

Now the version is once again hard-coded in the build script. This is an
unfortunate decoupling from and duplication of version information in
Git itself, and means that it's possible to forget to change the
version and deploy a revision of the application that is actually
different than the advertised version with no easy way of knowing that
this happened, but that's no different than most applications. Given how
much simpler an arrangement this presents, it's the favorable one, even
if it requires a bit more diligence and loses a little information.