Read version from classpath:version.txt
This change moves the application's version to
src/main/resources/version.txt instead of tracking it in the Gradle
build file. This allows for reliably reading and presenting the contents
of the version file regardless of whether the application was built via
Gradle or via IDEA. Previously, the /getVersion and /info endpoints
relied on application.properties being process via Gradle's
processResources task. Now everything works identically whether run from
Gradle or IDEA, and this also avoids the awkward problem of Spring and
Gradle competing for ${...}-formatted placeholder replacement.

Note that the version is still written to the jar manifest file, as this
is what Spring Boot reads from in order to replace the
${application.version}, or now ${application.formatted-version}
placeholder in banner.txt. See
https://docs.spring.io/spring-boot/docs/current/reference/htmlsingle/#boot-features-banner
for details.