Support a ~/.config/bisq.properties config file
If present, this file is loaded and provided to the Spring application
context as an additional source of properties, providing a more
convenient and persistent way to specify properties like the required
BitcoinAverage keypair. Specifying these properties via environment
variables remains supported as always.

Note that this file is called bisq.properties, and not, e.g.
bisq-pricenode.properties in anticipation that we may want to use this
same config file for similar convenient configuration of other Bisq node
types including exchange nodes themselves.