Handle exception if a provider is not available
In case a provider is not available we still want to deliver the data
of the other providers, so we catch a possible exception and leave
timestamp at 0. The Bisq app will check if the timestamp is in a
tolerance window and if it is too old it will show that the price is
not available.